Tested with background job that does not register against the registry
