+++ domains/certificates/Query.API/API.IntegrationTests/HoleTest.cs
@@ -1,33 +1,33 @@
-extern alias registryConnector;
-using System;
-using System.Threading.Tasks;
-using API.IntegrationTests.Factories;
-using API.IntegrationTests.Testcontainers;
-using FluentAssertions;
-using Xunit;
+//extern alias registryConnector;
+//using System;
+//using System.Threading.Tasks;
+//using API.IntegrationTests.Factories;
+//using API.IntegrationTests.Testcontainers;
+//using FluentAssertions;
+//using Xunit;
 
-namespace API.IntegrationTests;
+//namespace API.IntegrationTests;
 
-public class HoleTest : IClassFixture<RegistryConnectorApplicationFactory>, IClassFixture<ProjectOriginStack>
-{
-    private readonly RegistryConnectorApplicationFactory factory;
-    private readonly ProjectOriginStack projectOriginStack;
+//public class HoleTest : IClassFixture<RegistryConnectorApplicationFactory>, IClassFixture<ProjectOriginStack>
+//{
+//    private readonly RegistryConnectorApplicationFactory factory;
+//    private readonly ProjectOriginStack projectOriginStack;
 
-    public HoleTest(RegistryConnectorApplicationFactory factory, ProjectOriginStack projectOriginStack)
-    {
-        this.factory = factory;
-        this.projectOriginStack = projectOriginStack;
+//    public HoleTest(RegistryConnectorApplicationFactory factory, ProjectOriginStack projectOriginStack)
+//    {
+//        this.factory = factory;
+//        this.projectOriginStack = projectOriginStack;
 
-        factory.ProjectOriginOptions = projectOriginStack.Options;
-    }
+//        factory.ProjectOriginOptions = projectOriginStack.Options;
+//    }
 
-    [Fact]
-    public async Task Test1()
-    {
-        factory.Start();
+//    [Fact]
+//    public async Task Test1()
+//    {
+//        factory.Start();
 
-        await Task.Delay(TimeSpan.FromSeconds(10));
+//        await Task.Delay(TimeSpan.FromSeconds(70));
 
-        projectOriginStack.WalletUrl.Should().Be("http://127.0.0.1:7890/");
-    }
-}
+//        projectOriginStack.WalletUrl.Should().Be("http://127.0.0.1:7890/");
+//    }
+//}
